Capture output in `@optimize_or_load`
Running `@optimize_or_load` now captures the output from the underlying
call to `optimize` and stores it in the `.jld2` file alongside the
result. In a subsequent call to `@optimize_or_load`, print the captured
output. The captured output is limited to 2kB. If `optimize` produces
more output than that, the first and last kB ouf output are captured.

Added a `logfile` option to `@optimize_or_load` for situations where
the output is larger than 2kB but should be fully preserved.