Make `chi` functions not-in-place
The `chi!` functions previously used by GRAPE and Krotov are now
simply `chi` do not act in-place. This is more general and easier to
implement for the user, as it allows to use immutable structs for
states

Note that in extreme performance-critical situations, one could
still construct the χ-states in-place via a closure or functor.

See also
https://github.com/JuliaQuantumControl/QuantumControlBase.jl/pull/85