Make grad_J_a act not-in-place
This is much easier for a user to deal with. It really shouldn't make
any difference for performance, and in fact, may enhance it (the old
implementation was using `copyto!`). In situations where it would make a
difference, like for insanely large pulses, there's always the option of
using a non-allocating functor.