Add `AbstractOptimizationResult`
This allows for more seamless conversion between the result objects of
different methods. To convert between two result types, the first result
is converted to a Dict of field names to values, and then that dict is
converted to the target result type. This assumes that all result types
have a common set of field names, and for any field in a result that is
not in that common set, a custom convert method must be defined that
sets default values for those fields in the target result type.