In `make_chi`, introspect `J_T` to determine default `via`
The `via=:tau` option for `make_chi` requires that `J_T` takes a `tau`
keyword argument. For the default "automatic" keyword for `via`, this
should be checked explicitly via introspection. Otherwise, the naive
definition of a `J_T` function by a user that does not use `tau` even
though the functional is defined in terms of target states leads to a
hard-to-understand error.